backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
aiomysql>=0.2
PyMySQL>=1.1
pydantic>=2.5
pydantic-settings>=2.2
PyJWT>=2.8
bcrypt>=4.1
python-multipart>=0.0.9
python-dotenv>=1.0
torch>=1.7
torchvision>=0.8
opencv-python>=4.5
numpy>=1.21
requests>=2.25
PyYAML>=5.3
tqdm>=4.41
scipy>=1.4
pandas>=1.1
seaborn>=0.11
thop>=0.1.1
